requirements.txt
beautifulsoup4~=4.12.3
jupyter~=1.0.0
gpt4all~=2.7.0
langchain~=0.2.6
langchainhub~=0.1.20
langchain-community~=0.2.6
langchain-chroma~=0.1.2
unstructured[pdf]~=0.14.9
python-dotenv~=1.0.1
pdfminer.six==20231228
pillow~=10.4.0
s3fs~=2024.6.1
# S3FileLoader in langchain-community.document_loaders dynamically imports boto3
# s3fs requires aiobotocore. The latest aiobotocore, 2.13.1, requires botocore<1.34.132,>=1.34.70,
# so we have to ensure we use a compatible boto3
boto3==1.34.131
